Remove commonjs / umd builds Split typings and add some tests
