+++ rollup.config.js
@@ -1,17 +1,10 @@
 'use strict';
 
-import pkg from './package.json';
 import typescript from '@rollup/plugin-typescript';
 import { terser } from 'rollup-plugin-terser';
 import size from 'rollup-plugin-size';
 
-const plugins = [
-  typescript({ typescript: require('typescript') }),
-  terser(),
-  size(),
-]
-
-export default [{
+export default {
   input: {
     index: 'src/index.ts',
     combineReducers: 'src/combineReducers.ts',
@@ -27,25 +20,9 @@ export default [{
     format: 'esm',
     sourcemap: true,
   },
-  plugins,
-}, {
-  input: 'src/index.ts',
-  output: [{
-    file: pkg.main,
-    format: 'cjs',
-  }, {
-    file: 'lib/index.min.js',
-    format: 'esm',
-    sourcemap: true,
-  }],
-  plugins,
-}, {
-  input: 'src/index.ts',
-  output: {
-    file: pkg.browser,
-    format: 'umd',
-    name: 'store',
-    esModule: false
-  },
-  plugins,
-}]
+  plugins: [
+    typescript({ typescript: require('typescript') }),
+    terser(),
+    size(),
+  ]
+}
